upgraded vitepress to v1.0.0-alpha.29; made changes so the wiki runs without any problems; added fastbuild to exclude big files (specified in config.ts) while building; added run-script-os so that npm-scripts can be run on all os
