+++ .prettierrc.json
@@ -1,8 +1,17 @@
 {
-    "trailingComma": "es5",
-    "useTabs": true,
-    "tabWidth": 4,
-    "semi": false,
-    "singleQuote": true,
-    "printWidth": 80
+	"trailingComma": "es5",
+	"useTabs": true,
+	"tabWidth": 4,
+	"semi": false,
+	"singleQuote": true,
+	"printWidth": 80,
+	"embeddedLanguageFormatting": "off",
+	"overrides": [
+		{
+			"files": "*.md",
+			"options": {
+				"parser": "mdx"
+			}
+		}
+	]
 }
