+++ docs/.vitepress/theme/Components/Content/FolderView/FolderView.vue
@@ -1,8 +1,8 @@
 <template>
-	<div class="border-2 p-2 m-2 rounded-md border-gray">
-		<FolderViewChild :depth="-1" :name="'.'" :nodes="getData()">
-		</FolderViewChild>
-	</div>
+  <div class="border-2 p-2 m-2 rounded-md border-gray">
+    <FolderViewChild :depth="-1" :name="'.'" :nodes="getData()">
+    </FolderViewChild>
+  </div>
 </template>
 
 <script lang="js">
