+++ docs/.vitepress/theme/Components/Navigation/NavLink.vue
@@ -1,13 +1,9 @@
 <!--Volar complains here but everything is working fine-->
 <template>
-	<a
-		class="item flex flex-nowrap items-center px-2 rounded-md"
-		:class="[{ section: props.item.section }, props.item.color]"
-		v-bind="linkProps"
-		@click="onNavigation"
-	>
+	<a class="item flex flex-nowrap items-center px-2 rounded-md"
+		:class="[{ section: props.item.section }, props.item.color]" v-bind="linkProps" @click="onNavigation">
 		<span v-if="props.item.prefix != null" class="font-bold pr-2">{{
-			props.item.prefix
+				props.item.prefix
 		}}</span>
 
 		<div class="text-sm">{{ props.item.text }}</div>
@@ -25,9 +21,10 @@ import NavLinkTag from './NavLinkTag.vue'
 import ExternalIcon from '../Icons/ExternalIcon.vue'
 
 import { toRefs, watchEffect } from 'vue'
-import { useNavLink } from 'vitepress/dist/client/theme-default/composables/navLink'
+import { useNavLink } from '../../Composables/navLink'
 import { useIsMobile } from '../../Composables/isMobile'
 import { useSidebarState } from '../../Composables/sidebar'
+
 
 const emit = defineEmits(['change'])
 
@@ -63,9 +60,11 @@ watchEffect(() => {
 	width: 16px;
 	margin-left: 4px;
 }
+
 .item {
 	@apply transition-colors px-2 py-1 m-1;
 }
+
 .item:hover,
 .item.active {
 	@apply bg-true-gray-200;
@@ -74,6 +73,7 @@ watchEffect(() => {
 .item.active {
 	@apply text-purple-800;
 }
+
 .dark .item.active {
 	@apply text-purple-300;
 }
