+++ docs/.vitepress/theme/Utils/fetch.ts
@@ -1,8 +1,8 @@
-export function universalFetch(url: string, headers: any) {
+import fetch from 'node-fetch'
+
+export async function universalFetch(url: string, headers: any) {
 	if (import.meta.env.SSR) {
-		return import('node-fetch').then((nodeFetch) =>
-			(<any>nodeFetch)(url, headers)
-		)
+		return await fetch(url, headers)
 	}
 
 	return fetch(url, headers)
