+++ docs/.vitepress/theme/Utils/full-headers.ts
@@ -4,7 +4,7 @@ const { remove: removeDiacritics } = require('diacritics')
 const rControl = /[\u0000-\u001f]/g
 const rSpecial = /[\s~`!@#$%^&*()\-_+=[\]{}|\\;:"'<>,.?/]+/g
 
-const slugify = (str) => {
+const slugify = (str: any) => {
 	return (
 		removeDiacritics(str)
 			// Remove control characters
@@ -22,14 +22,22 @@ const slugify = (str) => {
 	)
 }
 
-function extractHeaderPlugin(md, include = ['h1', 'h2', 'h3']) {
-	md.renderer.rules.heading_open = (tokens, i, options, env, self) => {
+function extractHeaderPlugin(md: any, include = ['h1', 'h2', 'h3']) {
+	md.renderer.rules.heading_open = (
+		tokens: any,
+		i: any,
+		options: any,
+		env: any,
+		self: any
+	) => {
 		const token = tokens[i]
 		if (include.includes(token.tag)) {
 			const title = tokens[i + 1].content
-			const idAttr = token.attrs.find(([name]) => name === 'id')
+			const idAttr = token.attrs.find(
+				([name]: [string]): boolean => name === 'id'
+			)
 			const slug = idAttr && idAttr[1]
-			const data = md.__data
+			const data = md.__data || (md.__data = {})
 			const headers = data.headers || (data.headers = [])
 			headers.push({
 				level: parseInt(token.tag.slice(1), 10),
