+++ docs/concepts/contents.md
@@ -5,7 +5,7 @@ mentions:
     - Osaxely
 ---
 
-`contents.json` is a file that is *probably* used for the game to process the pack files more easily. It is *probably* intended for marketplace content creators and Mojang, it is not required to have this file in the pack for the pack to work properly.
+`contents.json` is a file that is _probably_ used for the game to process the pack files more easily. It is _probably_ intended for marketplace content creators and Mojang, it is not required to have this file in the pack for the pack to work properly.
 
 You will find there some instructions about the usage of this file.
 
@@ -18,50 +18,47 @@ Example:
 
 ```json
 {
-  "content": [
-    {
-      "path": "texts/en_US.lang"
-    },
-    {
-      "path": "contents.json"
-    },
-    {
-      "path": "manifest.json"
-    },
-    {
-      "path": "animations/my_animation.animation.json"
-    },
-    {
-      "path": "animation_controllers/my_ac.ac.json"
-    },
-    {
-      "path": "entity/my_entity.entity.json"
-    },
-    {
-      "path": "textures/textures_list.json"
-    },
-    {
-      "path": "textures/blocks/my_block.png"
-    }
-  ]
+	"content": [
+		{
+			"path": "texts/en_US.lang"
+		},
+		{
+			"path": "contents.json"
+		},
+		{
+			"path": "manifest.json"
+		},
+		{
+			"path": "animations/my_animation.animation.json"
+		},
+		{
+			"path": "animation_controllers/my_ac.ac.json"
+		},
+		{
+			"path": "entity/my_entity.entity.json"
+		},
+		{
+			"path": "textures/textures_list.json"
+		},
+		{
+			"path": "textures/blocks/my_block.png"
+		}
+	]
 }
 ```
 
 <FolderView
 	:paths="[
-
-	'RP/texts/en_US.lang',
+    'RP/texts/en_US.lang',
     'RP/manifest.json',
-	'RP/contents.json',
+    'RP/contents.json',
     'RP/animations/my_animation.animation.json',
-	'RP/animation_controllers/my_ac.ac.json',
-	'RP/entity/my_entity.entity.json',
-	'RP/textures/texture_list.json',
-	'RP/textures/blocks/my_block.png'
-
-
+    'RP/animation_controllers/my_ac.ac.json',
+    'RP/entity/my_entity.entity.json',
+    'RP/textures/texture_list.json',
+    'RP/textures/blocks/my_block.png'
 ]"
-></FolderView>
+> </FolderView>
 
 ## Automatizing the process
 
@@ -72,10 +69,11 @@ The `contents.json` file can be generated automatically by the game itself, it i
 ```json
 {}
 ```
+
 The file content will be automatically written next time the game is launched.
 
 ## Additional information
 
--	The automatic process can be achieved no matter what is the location of the pack (Development folders or normal folders).
--	Do not make multiple `contents.json` for subpacks, the file at the root of the pack is sufficient.
--	This file is not required for the addon to work properly.
+-   The automatic process can be achieved no matter what is the location of the pack (Development folders or normal folders).
+-   Do not make multiple `contents.json` for subpacks, the file at the root of the pack is sufficient.
+-   This file is not required for the addon to work properly.
