+++ docs/documentation/pack-structure.md
@@ -4,132 +4,101 @@ show_toc: false
 ---
 
 <FolderView :paths="[
-
-'BP/manifest.json',
-'BP/pack_icon.png',
-
-'BP/animations/example.animation.json',
-'BP/animation_controllers/example.ac.json',
-
-'BP/blocks/example.block.json',
-
-'BP/biomes/example.biome.json',
-
-'BP/entities/example.se.json',
-
-'BP/features/example.feature.json',
-'BP/feature_rules/example.rule.json',
-
-'BP/functions/example.mcfunction',
-'BP/functions/tick.json',
-
-'BP/items/example.item.json',
-
-'BP/loot_tables/example.loot.json',
-
-'BP/recipes/example.recipe.json',
-
-'BP/scripts/client/exampleClient.js',
-'BP/scripts/server/exampleServer.js',
-'BP/scripts/gametests/exampleTest.js',
-
-'BP/spawn_rules/example.spawn.json',
-
-'BP/texts/languages.json',
-'BP/texts/\*.lang',
-
-'BP/trading/example.trade.json',
-'BP/trading/economy_trades/example.trade.json',
-
-'BP/structures/example.mcstructure',
-
-'RP/manifest.json',
-'RP/pack_icon.png',
-'RP/biomes_client.json',
-'RP/blocks.json',
-'RP/sounds.json',
-'RP/contents.json',
-
-'RP/animation_controllers/example.ac.json',
-'RP/animations/example.animation.json',
-
-'RP/attachables/example.attachable.json',
-
-'RP/entity/example.ce.json',
-
-'RP/fogs/example_fog_setting.json',
-
-'RP/items/example.item.json',
-
-'RP/materials/example.material',
-                    
-'RP/models/entity/example.geo.json',
-'RP/models/blocks/example.geo.json',
-
-'RP/particles/example.particle.json',
-
-'RP/render_controllers/example.rc.json',
-
-'RP/sounds/example.wav',
-'RP/sounds/example.ogg',
-'RP/sounds/example.mp3',
-'RP/sounds/example.fsb',
-'RP/sounds/sound_definitions.json',
-'RP/sounds/music_definitions.json',
-
-'RP/texts/languages.json',
-'RP/texts/language_names.json',
-'RP/texts/bg_BG.lang',
-'RP/texts/cs_CZ.lang',
-'RP/texts/da_DK.lang',
-'RP/texts/de_DE.lang',
-'RP/texts/el_GR.lang',
-'RP/texts/en_GB.lang',
-'RP/texts/en_US.lang',
-'RP/texts/es_ES.lang',
-'RP/texts/es_MX.lang',
-'RP/texts/fi_FI.lang',
-'RP/texts/fr_CA.lang',
-'RP/texts/fr_FR.lang',
-'RP/texts/hu_HU.lang',
-'RP/texts/id_ID.lang',
-'RP/texts/it_IT.lang',
-'RP/texts/ja_JP.lang',
-'RP/texts/ko_KR.lang',
-'RP/texts/nb_NO.lang',
-'RP/texts/nl_NL.lang',
-'RP/texts/pl_PL.lang',
-'RP/texts/pt_BR.lang',
-'RP/texts/pt_PR.lang',
-'RP/texts/ru_RU.lang',
-'RP/texts/sk_SK.lang',
-'RP/texts/sv_SE.lang',
-'RP/texts/tr_TR.lang',
-'RP/texts/uk_UA.lang',
-'RP/texts/zh_CN.lang',
-'RP/texts/zh_TW.lang',
-'RP/texts/zh_TW.lang',
-'RP/texts/ja_JP/font/glyph_2E.png',
-'RP/texts/ja_JP/font/\*.png',
-'RP/texts/zh_TW/font/glyph_2E.png',
-'RP/texts/zh_TW/font/\*.png',
-
-'RP/textures/item_texture.json',
-'RP/textures/terrain_texture.json',
-'RP/textures/flipbook_textures.json',
-'RP/textures/texture_list.json',
-
-'RP/textures/environment/overworld_cubemap/cubemap_0.png',
-'RP/textures/environment/overworld_cubemap/cubemap_1.png',
-'RP/textures/environment/overworld_cubemap/cubemap_2.png',
-'RP/textures/environment/overworld_cubemap/cubemap_3.png',
-'RP/textures/environment/overworld_cubemap/cubemap_4.png',
-'RP/textures/environment/overworld_cubemap/cubemap_5.png',
-
-'RP/textures/blocks/example.png',
-'RP/textures/entity/example.png',
-'RP/textures/items/example.png',
-'RP/textures/particle/example.png',
-
-'RP/ui/\*.json'
+    'BP/manifest.json',
+    'BP/pack_icon.png',
+    'BP/animations/example.animation.json',
+    'BP/animation_controllers/example.ac.json',
+    'BP/blocks/example.block.json',
+    'BP/biomes/example.biome.json',
+    'BP/entities/example.se.json',
+    'BP/features/example.feature.json',
+    'BP/feature_rules/example.rule.json',
+    'BP/functions/example.mcfunction',
+    'BP/functions/tick.json',
+    'BP/items/example.item.json',
+    'BP/loot_tables/example.loot.json',
+    'BP/recipes/example.recipe.json',
+    'BP/scripts/client/exampleClient.js',
+    'BP/scripts/server/exampleServer.js',
+    'BP/scripts/gametests/exampleTest.js',
+    'BP/spawn_rules/example.spawn.json',
+    'BP/texts/languages.json',
+    'BP/texts/\*.lang',
+    'BP/trading/example.trade.json',
+    'BP/trading/economy_trades/example.trade.json',
+    'BP/structures/example.mcstructure',
+    'RP/manifest.json',
+    'RP/pack_icon.png',
+    'RP/biomes_client.json',
+    'RP/blocks.json',
+    'RP/sounds.json',
+    'RP/contents.json',
+    'RP/animation_controllers/example.ac.json',
+    'RP/animations/example.animation.json',
+    'RP/attachables/example.attachable.json',
+    'RP/entity/example.ce.json',
+    'RP/fogs/example_fog_setting.json',
+    'RP/items/example.item.json',
+    'RP/materials/example.material',              
+    'RP/models/entity/example.geo.json',
+    'RP/models/blocks/example.geo.json',
+    'RP/particles/example.particle.json',
+    'RP/render_controllers/example.rc.json',
+    'RP/sounds/example.wav',
+    'RP/sounds/example.ogg',
+    'RP/sounds/example.mp3',
+    'RP/sounds/example.fsb',
+    'RP/sounds/sound_definitions.json',
+    'RP/sounds/music_definitions.json',
+    'RP/texts/languages.json',
+    'RP/texts/language_names.json',
+    'RP/texts/bg_BG.lang',
+    'RP/texts/cs_CZ.lang',
+    'RP/texts/da_DK.lang',
+    'RP/texts/de_DE.lang',
+    'RP/texts/el_GR.lang',
+    'RP/texts/en_GB.lang',
+    'RP/texts/en_US.lang',
+    'RP/texts/es_ES.lang',
+    'RP/texts/es_MX.lang',
+    'RP/texts/fi_FI.lang',
+    'RP/texts/fr_CA.lang',
+    'RP/texts/fr_FR.lang',
+    'RP/texts/hu_HU.lang',
+    'RP/texts/id_ID.lang',
+    'RP/texts/it_IT.lang',
+    'RP/texts/ja_JP.lang',
+    'RP/texts/ko_KR.lang',
+    'RP/texts/nb_NO.lang',
+    'RP/texts/nl_NL.lang',
+    'RP/texts/pl_PL.lang',
+    'RP/texts/pt_BR.lang',
+    'RP/texts/pt_PR.lang',
+    'RP/texts/ru_RU.lang',
+    'RP/texts/sk_SK.lang',
+    'RP/texts/sv_SE.lang',
+    'RP/texts/tr_TR.lang',
+    'RP/texts/uk_UA.lang',
+    'RP/texts/zh_CN.lang',
+    'RP/texts/zh_TW.lang',
+    'RP/texts/zh_TW.lang',
+    'RP/texts/ja_JP/font/glyph_2E.png',
+    'RP/texts/ja_JP/font/\*.png',
+    'RP/texts/zh_TW/font/glyph_2E.png',
+    'RP/texts/zh_TW/font/\*.png',
+    'RP/textures/item_texture.json',
+    'RP/textures/terrain_texture.json',
+    'RP/textures/flipbook_textures.json',
+    'RP/textures/texture_list.json',
+    'RP/textures/environment/overworld_cubemap/cubemap_0.png',
+    'RP/textures/environment/overworld_cubemap/cubemap_1.png',
+    'RP/textures/environment/overworld_cubemap/cubemap_2.png',
+    'RP/textures/environment/overworld_cubemap/cubemap_3.png',
+    'RP/textures/environment/overworld_cubemap/cubemap_4.png',
+    'RP/textures/environment/overworld_cubemap/cubemap_5.png',
+    'RP/textures/blocks/example.png',
+    'RP/textures/entity/example.png',
+    'RP/textures/items/example.png',
+    'RP/textures/particle/example.png',
+    'RP/ui/\*.json'
 ]"></FolderView>
