+++ docs/privacy.md
@@ -5,48 +5,51 @@ show_toc: true
 show_contributors: false
 ---
 
-[wiki.bedrock.dev](/) is created and maintained by the Bedrock-OSS organization. This privacy policy will explain how our organization uses the personal data we collect from you when you use our website. By using the wiki.bedrock.dev site, you agree to this policy. 
+[wiki.bedrock.dev](/) is created and maintained by the Bedrock-OSS organization. This privacy policy will explain how our organization uses the personal data we collect from you when you use our website. By using the wiki.bedrock.dev site, you agree to this policy.
 
 ## What data do we collect?
 
 Bedrock-OSS collects the following information:
-- First two fields of your IP address (e.g. 8.8.0.0 instead of 8.8.8.8)
-- Date and time
-- Title and URL of the page visited
-- URL of the referring page
-- Screen resolution
-- Timezone
-- Files downloaded from the site
-- Links to an outside domain clicked
-- Time taken for pages to load
-- Approximate location and region
-- Browser language
-- Browser User-Agent
+
+-   First two fields of your IP address (e.g. 8.8.0.0 instead of 8.8.8.8)
+-   Date and time
+-   Title and URL of the page visited
+-   URL of the referring page
+-   Screen resolution
+-   Timezone
+-   Files downloaded from the site
+-   Links to an outside domain clicked
+-   Time taken for pages to load
+-   Approximate location and region
+-   Browser language
+-   Browser User-Agent
 
 ## How do we collect your data?
 
 You directly provide us with all the data we collect. We collect data when you:
-- Open any page on the wiki.bedrock.dev domain
-- Click any external link
+
+-   Open any page on the wiki.bedrock.dev domain
+-   Click any external link
 
 ## How will we use your data?
 
-Bedrock-OSS collects your data for the purpose of improving the site and evaluating the popularity of different pages and the site as a whole over time.  
-- Location, region and language are recorded to compare the use of the site in different languages
-- Page visited is used to evaluate the popularity of pages and influence future work on the site
-- Screen resolution is used to help prioritize development for different target devices
+Bedrock-OSS collects your data for the purpose of improving the site and evaluating the popularity of different pages and the site as a whole over time.
 
-Bedrock-OSS does not share your data with any other entities except in broad numbers, such as the total number of views a page has recieved in a week. 
+-   Location, region and language are recorded to compare the use of the site in different languages
+-   Page visited is used to evaluate the popularity of pages and influence future work on the site
+-   Screen resolution is used to help prioritize development for different target devices
+
+Bedrock-OSS does not share your data with any other entities except in broad numbers, such as the total number of views a page has recieved in a week.
 
 ## How do we store your data?
 
-Bedrock-OSS stores all analytics data on a private server accessible only to organization leaders. 
+Bedrock-OSS stores all analytics data on a private server accessible only to organization leaders.
 
-Your data remains on our servers for 6 months, after which it is erased. 
+Your data remains on our servers for 6 months, after which it is erased.
 
 ## What are your data protection rights?
 
-At Bedrock-OSS, we would like to ensure you are fully aware of your data protection rights. 
+At Bedrock-OSS, we would like to ensure you are fully aware of your data protection rights.
 
 In the collection of your data, we do not separate individual users, but store each request separately.
 
@@ -57,11 +60,11 @@ In the collection of your data, we do not separate individual users, but store e
 
 ## How do we use cookies?
 
-Our site uses cookies to improve user experience, such as storing the user's preferred color scheme. We do not use any cookies for tracking or analytics, unless you have prevented the site from tracking you above, in which case we use a cookie to remember your choice. 
+Our site uses cookies to improve user experience, such as storing the user's preferred color scheme. We do not use any cookies for tracking or analytics, unless you have prevented the site from tracking you above, in which case we use a cookie to remember your choice.
 
 ## Privacy Policies of other websites
 
-Our site contains links to other sites, which operate under different policies. We are not responsible for the privacy policies of other sites. 
+Our site contains links to other sites, which operate under different policies. We are not responsible for the privacy policies of other sites.
 
 ## Changes to our Privacy Policy
 
@@ -69,4 +72,4 @@ We try to keep our privacy policy updated. This page was last updated on 18 Janu
 
 ## Contact us
 
-To contact us, join our [discord server](/discord) or send an email to [SirLich.business@gmail.com](mailto:SirLich.business@gmail.com). 
+To contact us, join our [discord server](/discord) or send an email to [SirLich.business@gmail.com](mailto:SirLich.business@gmail.com).
