+++ package.json
@@ -3,33 +3,42 @@
 	"version": "1.0.0",
 	"main": "index.js",
 	"license": "MIT",
-	"devDependencies": {
-		"@docsearch/css": "^1.0.0-alpha.28",
-		"@docsearch/js": "^1.0.0-alpha.28",
-		"@types/node-fetch": "^2.5.11",
-		"diacritics": "^1.3.0",
-		"gray-matter": "^4.0.3",
-		"node-fetch": "^2.6.1",
-		"sass": "^1.35.2",
-		"vite-plugin-pwa": "^0.8.2",
-		"vite-plugin-windicss": "^1.8.3",
-		"vitepress": "^0.22.3",
-		"windicss": "^3.5.1"
-	},
 	"scripts": {
-		"dev": "vitepress dev docs",
-		"build": "node --max-old-space-size=8192 node_modules/vitepress/bin/vitepress.js build docs",
+		"dev": "run-script-os",
+		"dev:win32": "SET fastBuild=false && vitepress dev docs",
+		"dev:darwin:linux": "fastBuild=false && vitepress dev docs",
+		"build": "run-script-os",
+		"build:win32": "SET fastBuild=false && node --max-old-space-size=16384 node_modules/vitepress/bin/vitepress.js build docs",
+		"build:darwin:linux": "fastBuild=false && node --max-old-space-size=16384 node_modules/vitepress/bin/vitepress.js build docs",
+		"fastbuild": "run-script-os",
+		"fastbuild:win32": "SET fastBuild=true && node --max-old-space-size=8192 node_modules/vitepress/bin/vitepress.js build docs",
+		"fastbuild:darwin:linux": "fastBuild=true && node --max-old-space-size=8192 node_modules/vitepress/bin/vitepress.js build docs",
 		"serve": "vitepress serve docs"
 	},
 	"dependencies": {
-		"@types/prismjs": "^1.16.6",
-		"@vueuse/core": "^5.1.3",
-		"markdown-it": "^12.1.0",
-		"molang": "^1.13.3",
-		"molangjs": "^1.4.1",
-		"prismjs": "^1.25.0",
-		"vite": "^2.6.14",
-		"vue": "^3.2.22",
+		"@types/prismjs": "^1.26.0",
+		"@vueuse/core": "^9.5.0",
+		"markdown-it": "^13.0.1",
+		"molang": "^2.0.1",
+		"molangjs": "^1.5.2",
+		"prismjs": "^1.29.0",
+		"vite": "^3.2.4",
 		"vue-prism-editor": "^2.0.0-alpha.2"
+	},
+	"devDependencies": {
+		"@docsearch/css": "^3.3.0",
+		"@docsearch/js": "^3.3.0",
+		"@types/node": "^18.11.9",
+		"@types/node-fetch": "^2.5.11",
+		"diacritics": "^1.3.0",
+		"gray-matter": "^4.0.3",
+		"node-fetch": "^2.6.7",
+		"run-script-os": "^1.1.6",
+		"sass": "^1.56.1",
+		"vite-plugin-pwa": "^0.13.3",
+		"vite-plugin-windicss": "^1.8.8",
+		"vitepress": "^1.0.0-alpha.29",
+		"vue": "^3.2.45",
+		"windicss": "^3.5.6"
 	}
 }
