+++ tsconfig.json
@@ -10,7 +10,7 @@
 		"esModuleInterop": true,
 		"allowJs": true,
 		"lib": ["esnext", "dom"],
-		"types": ["vite/client"]
+		"types": ["vite/client", "node"]
 	},
 	"include": [
 		"docs/.vitepress/**/*.ts",
